moving setup_logging to logging module
